+++ frb_codegen/src/library/codegen/parser/hir/hierarchical/mod.rs
@@ -9,7 +9,6 @@ pub(crate) mod mirror_ident;
 pub(crate) mod module;
 pub(crate) mod pack;
 pub(crate) mod struct_or_enum;
-mod third_party_override_transformer;
 pub(crate) mod visibility;
 
 pub(crate) fn parse(
